fix Fungsi export laporan

+++ database/migrations/2025_11_18_122358_create_responden_ikm_pembinaans_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('id_biodata');
             $table->string('kd_unsurikmpembinaan')->nullable();
-            $table->integer('skor');
+            $table->decimal('skor', 5, 2);
             $table->timestamps();
 
             $table->foreignId('narasumber_id')->nullable()->constrained('narasumbers')->onDelete('set null');
